refactor: remove DaemonManaged abstraction and dead non-daemon code
The agent always runs in daemon mode — there is no other Host
implementation. Remove DaemonManaged() and AutoCreatePR() from the Host
interface, delete the worker-side auto-merge system (auto_merge.go),
and simplify handleCreatePR/handlePushBranch to always reject since the
daemon workflow manages PR creation and pushing.
